requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
multidict==4.6.1
pyee==6.0.0
requests==2.22.0
six==1.13.0
slackclient==2.3.1
slackeventsapi==2.1.0
urllib3==1.25.7
Werkzeug==0.16.0
yarl==1.3.0
